requirements.txt
# Backend Framework
flask==3.0.0
flask-cors==4.0.0

# AI/ML APIs
google-generativeai==0.3.2

# PDF Processing
pypdf2==3.0.1

# HTTP requests
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1

